Introduce the concept of local-only keys.
Values for local keys will not be distributed (will live only on the node which received them)
and will be stored on the local filesystem. The purpose is allowing to use a node
to store and retrieve big data that wouldn't otherwise fit in memory.
For example large files or anyway data which can be transfered over the network in chunks but can't
be fully loaded in memory (hence cached)